refactor(artifacts): Move artifact binding to orca-core (#2021)
In order to use the artifact binding functionality in the image bakery,
we'll need this functionality to be available in orca-core so it can
be called from orca-bakery.
